# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend para la optimización de producción de una panadería usando LINGO.",
        "title": "LINGO Bakery API",
        "contact": {
            "name": "Soporte Técnico"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/ingredients": {
            "get": {
                "description": "Obtiene la lista completa de ingredientes registrados en el sistema",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Listar ingredientes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Ingredient"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Registra un nuevo insumo básico en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Crear ingrediente",
                "parameters": [
                    {
                        "description": "Datos del ingrediente",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    }
                }
            }
        },
        "/ingredients/{id}": {
            "get": {
                "description": "Obtiene los detalles de un ingrediente específico",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ingredients"
                ],
                "summary": "Obtener ingrediente",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID del ingrediente",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Ingredient"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/optimize": {
            "post": {
                "description": "Encola un nuevo trabajo para que LINGO calcule la producción óptima",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Optimization"
                ],
                "summary": "Iniciar optimización",
                "parameters": [
                    {
                        "description": "IDs de Stock y Recursos",
                        "name": "output",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Optimization"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "job_id, status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/optimize/{job_id}": {
            "get": {
                "description": "Obtiene el estado (pending, processing, done, error) de un trabajo por su UUID y su ID de base de datos",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Optimization"
                ],
                "summary": "Consultar estado de Job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UUID del trabajo",
                        "name": "job_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "job_id, status, id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/results/{id}": {
            "get": {
                "description": "Obtiene las cantidades óptimas a producir y ganancias esperadas",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Optimization"
                ],
                "summary": "Obtener resultados de optimización",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID de la optimización (Postgres)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Optimization"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Ingredient": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "unit": {
                    "description": "kg, litros, unidades, etc.",
                    "type": "string"
                },
                "unit_cost": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Machine": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.OperationalResource": {
            "type": "object",
            "properties": {
                "available": {
                    "description": "DISP(R)",
                    "type": "number"
                },
                "cost_per_unit": {
                    "description": "CR(R)",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "resource_id": {
                    "type": "integer"
                }
            }
        },
        "models.Optimization": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "job_id": {
                    "type": "string"
                },
                "max_production": {
                    "description": "M",
                    "type": "number"
                },
                "min_variety": {
                    "description": "PRO",
                    "type": "integer"
                },
                "resource": {
                    "$ref": "#/definitions/models.Resource"
                },
                "resource_id": {
                    "type": "integer"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OptimizationResult"
                    }
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.OptimizationStatus"
                },
                "stock": {
                    "$ref": "#/definitions/models.Stock"
                },
                "stock_id": {
                    "type": "integer"
                },
                "total_profit": {
                    "type": "number"
                }
            }
        },
        "models.OptimizationResult": {
            "type": "object",
            "properties": {
                "batch_active": {
                    "description": "Y(I)",
                    "type": "number"
                },
                "expected_profit": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "optimization_id": {
                    "type": "integer"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity_to_produce": {
                    "description": "X(I)",
                    "type": "number"
                },
                "variety_flag": {
                    "description": "W(I)",
                    "type": "number"
                }
            }
        },
        "models.OptimizationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "done",
                "error",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusDone",
                "StatusError",
                "StatusCancelled"
            ]
        },
        "models.Product": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "demand": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductIngredient"
                    }
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductMachine"
                    }
                },
                "max_batch": {
                    "type": "number"
                },
                "min_batch": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "operational_resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ProductOperationalResource"
                    }
                },
                "sale_price": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ProductIngredient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/models.Ingredient"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "number"
                }
            }
        },
        "models.ProductMachine": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/models.Machine"
                },
                "machine_id": {
                    "type": "integer"
                },
                "minutes_per_unit": {
                    "type": "number"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "models.ProductOperationalResource": {
            "type": "object",
            "properties": {
                "consumption_per_batch": {
                    "description": "CM(I,R)",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "operational_resource": {
                    "$ref": "#/definitions/models.OperationalResource"
                },
                "operational_resource_id": {
                    "type": "integer"
                },
                "product_id": {
                    "type": "integer"
                }
            }
        },
        "models.Resource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "machines": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ResourceMachine"
                    }
                },
                "name": {
                    "type": "string"
                },
                "operational_resources": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OperationalResource"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.ResourceMachine": {
            "type": "object",
            "properties": {
                "hours_available": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "machine": {
                    "$ref": "#/definitions/models.Machine"
                },
                "machine_id": {
                    "type": "integer"
                },
                "resource_id": {
                    "type": "integer"
                }
            }
        },
        "models.Stock": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.StockIngredient"
                    }
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.StockIngredient": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "ingredient": {
                    "$ref": "#/definitions/models.Ingredient"
                },
                "ingredient_id": {
                    "type": "integer"
                },
                "quantity_available": {
                    "type": "number"
                },
                "stock_id": {
                    "type": "integer"
                }
            }
        }
    }
}